class/file cleanup
